requirements.txt
python-socketio[asyncio]
eventlet

